Refactor validation
